Move grouping manager into grouping dir
