+++ src/tree/azure/AzureResourceTreeDataProviderBase.ts
@@ -12,7 +12,7 @@ import { BranchDataItemCache } from '../BranchDataItemCache';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
 import { ResourceTreeDataProviderBase } from '../ResourceTreeDataProviderBase';
 import { TreeItemStateStore } from '../TreeItemState';
-import { AzureResourceGroupingManager } from './AzureResourceGroupingManager';
+import { AzureResourceGroupingManager } from './grouping/AzureResourceGroupingManager';
 import { GroupingItem } from './grouping/GroupingItem';
 
 export abstract class AzureResourceTreeDataProviderBase extends ResourceTreeDataProviderBase {
