+++ src/tree/azure/FocusViewTreeDataProvider.ts
@@ -14,8 +14,8 @@ import { settingUtils } from '../../utils/settingUtils';
 import { BranchDataItemCache } from '../BranchDataItemCache';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
 import { TreeItemStateStore } from '../TreeItemState';
-import { AzureResourceGroupingManager } from './AzureResourceGroupingManager';
 import { AzureResourceTreeDataProviderBase } from './AzureResourceTreeDataProviderBase';
+import { AzureResourceGroupingManager } from './grouping/AzureResourceGroupingManager';
 import { GroupingItem } from './grouping/GroupingItem';
 import { LocationGroupingItem } from './grouping/LocationGroupingItem';
 import { ResourceGroupGroupingItem } from './grouping/ResourceGroupGroupingItem';
