+++ src/tree/azure/registerFocusTree.ts
@@ -10,13 +10,13 @@ import { AzureResourceProviderManager } from '../../api/ResourceProviderManagers
 import { ext } from '../../extensionVariables';
 import { localize } from '../../utils/localize';
 import { BranchDataItemCache } from '../BranchDataItemCache';
-import { createTreeView } from '../createTreeView';
 import { ResourceGroupsItem } from '../ResourceGroupsItem';
+import { createTreeView } from '../createTreeView';
 import { wrapTreeForVSCode } from '../wrapTreeForVSCode';
 import { AzureResourceBranchDataProviderManager } from './AzureResourceBranchDataProviderManager';
-import { AzureResourceGroupingManager } from './AzureResourceGroupingManager';
 import { createResourceItemFactory } from './AzureResourceItem';
 import { FocusViewTreeDataProvider } from './FocusViewTreeDataProvider';
+import { AzureResourceGroupingManager } from './grouping/AzureResourceGroupingManager';
 import { GroupingItemFactory } from './grouping/GroupingItemFactory';
 
 interface RegisterAzureTreeOptions {
